feat(nav): implement stack navigation boilerplate
