+++ index.js
@@ -3,7 +3,7 @@
  */
 
 import {AppRegistry} from 'react-native';
-import App from './src/navigation/AppNavigator';
+import App from './src/navigation/StackNavigator';
 import {name as appName} from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
