--- src/navigation/AppNavigator.tsx
@@ -1,22 +1,0 @@
-import React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Home from '@/screens/Stack/Home'; 
-
-export default function MainNavigator() {
-
-  const Stack = createNativeStackNavigator();
-
-  return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen
-          name="Home"
-          component={Home}
-          options={{ headerShown: true }} // Hide the header for the Home screen
-        />
-        {/* Add more screens here as needed */}
-      </Stack.Navigator>
-    </NavigationContainer>
-  );
-}
